+++ lib/screens.dart
@@ -3,6 +3,7 @@ library screens;
 
 import 'package:flutter/material.dart';
 
+export 'src/screens/edit_point_screen.dart';
 export 'src/screens/error_screen.dart';
 export 'src/screens/loading_screen.dart';
 export 'src/screens/select_enum.dart';